virgl_renderer: Bump minimum virglrenderer version requirement to 1.0
Release of virglrenderer v1.0 stabilized lots of APIs and features.
In particular Venus became stable. Bump minimum virglrenderer version
requirement to 1.0.

This change removes vast majority of virgl_renderer_next usage. The
only unstable virglrenderer API that continues to be used is the optional
fence exporting feature.

The virglrenderer C binding was out of sync. The write_context_fence()
and virgl_renderer_context_create_fence() callback ABIs changed a year
ago. The virgl_renderer_export_fence() ABI was changed recently.
This all is fixed by this commit since the binding is regenerated.

TEST=none
Link: https://gitlab.freedesktop.org/virgl/virglrenderer/-/commit/9c9d55ecbe502e0d71ee99529709efcca69aac42
Link: https://gitlab.freedesktop.org/virgl/virglrenderer/-/commit/a5346dec8e848ca5de0e1a8d4bc03fcf988924d3
Link: https://gitlab.freedesktop.org/virgl/virglrenderer/-/commit/f8383c47991f7f13149e097346ce0c3f89802f3e
Suggested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4890315
Commit-Queue: Yiwei Zhang <zzyiwei@chromium.org>

+++ devices/src/virtio/gpu/virtio_gpu.rs
@@ -789,7 +789,7 @@ impl VirtioGpu {
     }
 
     /// If supported, export the fence with the given `fence_id` to a file.
-    pub fn export_fence(&self, fence_id: u32) -> ResourceResponse {
+    pub fn export_fence(&self, fence_id: u64) -> ResourceResponse {
         match self.rutabaga.export_fence(fence_id) {
             Ok(handle) => ResourceResponse::Resource(ResourceInfo::Fence {
                 handle: to_safe_descriptor(handle.os_handle),
